requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
Django==4.1.9
django-cors-headers==4.0.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pycparser==2.21
pymongo==3.11.0
pytz==2023.3
sqlparse==0.2.4
